Add helper triangle iterators to several widget modules
Includes triangle iterators for:

- The border of the BordererdRectangle widget
- PointPath
- Oval
- Polygon
- Rectangle

As well as a new `triangles::from_quad` function.

Update render module to use these accordingly
